fix(LottoService): WinningLotto 클래스에서 LottoInfo 클래스로 수정
LottoInfo
- 당첨번호, 보너스번호, 구매수량을 변수로 가집니다.
