+++ src/main/java/lotto/service/LottoService.java
@@ -3,17 +3,19 @@ package lotto.service;
 import camp.nextstep.edu.missionutils.Randoms;
 import lotto.model.Lotto;
 import lotto.dto.LottoRequestDto;
-import lotto.model.WinningLotto;
+import lotto.model.LottoInfo;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 public class LottoService {
-    public static WinningLotto createWinningLotto(LottoRequestDto lottoRequestDto) {
-        int bonusNumber = Integer.parseInt(lottoRequestDto.getBonusNumber());
+    public static LottoInfo createLottoInfo(LottoRequestDto lottoRequestDto) {
         List<Integer> lottoNumbers = parseLottoNumbers(lottoRequestDto.getLottoNumbers());
-        return new WinningLotto(lottoNumbers, bonusNumber);
+        int bonusNumber = Integer.parseInt(lottoRequestDto.getBonusNumber());
+        int amounts = parsePurchaseAmount(lottoRequestDto.getPurchaseAmount());
+
+        return new LottoInfo(lottoNumbers, bonusNumber, amounts);
     }
 
     public static Lotto createLotto() {
@@ -29,5 +31,9 @@ public class LottoService {
             lottoNumbers.add(Integer.parseInt(number));
         }
         return lottoNumbers;
+    }
+
+    private static int parsePurchaseAmount(String purchaseAmount) {
+        return Integer.parseInt(purchaseAmount) / 1000;
     }
 }
